feat(aci): Display all ongoing issues on error monitor page
